Split SDK plugin implementations
